Replaced - (unsigned int) count by - (NSUInteger) count in all frameworks to prevent return type conflict with GNUstep Foundation classes.
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileUI@6128 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Headers/ETStyleGroup.h
@@ -64,7 +64,7 @@ shared instance. Style groups can be used as shared style objects too though.  *
 
 - (BOOL) isOrdered;
 - (BOOL) isEmpty;
-- (unsigned int) count;
+- (NSUInteger) count;
 - (NSArray *) contentArray;
 - (void) addObject: (id)anObject;
 - (void) insertObject: (id)anObject atIndex: (unsigned int)anIndex;
